% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_event_api.R
\docType{class}
\name{ActivityEventApi}
\alias{ActivityEventApi}
\title{ActivityEvent operations}
\format{
An \code{R6Class} generator object
}
\description{
LAMP.ActivityEvent
}
\section{Methods}{

\describe{
\strong{ allByParticipant } \emph{ Get all activity events for a participant. }
Get the set of all activity events produced by a given participant,  by identifier.

\itemize{
\item \emph{ @param } participant.id character
\item \emph{ @param } origin character
\item \emph{ @param } from numeric
\item \emph{ @param } to numeric
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ allByResearcher } \emph{ Get all activity events for a researcher by participant. }
Get the set of all activity events produced by participants of any  study conducted by a researcher, by researcher identifier.

\itemize{
\item \emph{ @param } researcher.id character
\item \emph{ @param } origin character
\item \emph{ @param } from numeric
\item \emph{ @param } to numeric
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ allByStudy } \emph{ Get all activity events for a study by participant. }
Get the set of all activity events produced by participants of a  single study, by study identifier.

\itemize{
\item \emph{ @param } study.id character
\item \emph{ @param } origin character
\item \emph{ @param } from numeric
\item \emph{ @param } to numeric
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ create } \emph{ Create a new ActivityEvent for the given Participant. }
Create a new ActivityEvent for the given Participant.

\itemize{
\item \emph{ @param } participant.id character
\item \emph{ @param } activity.event \link{ActivityEvent}


\item status code : 200 | 200 Success

\item return type : character 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ delete } \emph{ Delete a ActivityEvent. }
Delete a ActivityEvent.

\itemize{
\item \emph{ @param } participant.id character
\item \emph{ @param } origin character
\item \emph{ @param } from numeric
\item \emph{ @param } to numeric
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : character 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  allByParticipant  ####################

library(LAMP)
var.participant.id <- 'participant.id_example' # character | 
var.origin <- 'origin_example' # character | 
var.from <- 3.4 # numeric | 
var.to <- 3.4 # numeric | 

#Get all activity events for a participant.
result <- LAMP.ActivityEvent$allByParticipant(var.participant.id, 
                                             origin=var.origin, 
                                             from=var.from, 
                                             to=var.to)


####################  allByResearcher  ####################

library(LAMP)
var.researcher.id <- 'researcher.id_example' # character | 
var.origin <- 'origin_example' # character | 
var.from <- 3.4 # numeric | 
var.to <- 3.4 # numeric | 

#Get all activity events for a researcher by participant.
result <- LAMP.ActivityEvent$allByResearcher(var.participant.id, 
                                            origin=var.origin, 
                                            from=var.from, 
                                            to=var.to)

####################  allByStudy  ####################

library(LAMP)
var.study.id <- 'study.id_example' # character | 
var.origin <- 'origin_example' # character | 
var.from <- 3.4 # numeric | 
var.to <- 3.4 # numeric | 

#Get all activity events for a study by participant.
result <- LAMP.ActivityEvent$allByStudy(var.study.id, origin=var.origin, from=var.from, to=var.to)


####################  create  ####################

library(LAMP)
var.participant.id <- 'participant.id_example' # character | 
var.activity.event <- ActivityEvent$new() # ActivityEvent | 

#Create a new ActivityEvent for the given Participant.
result <- LAMP.ActivityEvent$create(var.participant.id, var.activity.event)


####################  delete  ####################

library(LAMP)
var.participant.id <- 'participant.id_example' # character | 
var.origin <- 'origin_example' # character | 
var.from <- 3.4 # numeric | 
var.to <- 3.4 # numeric | 

#Delete a ActivityEvent.
result <- LAMP.ActivityEvent$delete(var.participant.id, origin=var.origin, from=var.from, to=var.to)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ActivityEventApi$new()}}
\item \href{#method-allByParticipant}{\code{ActivityEventApi$allByParticipant()}}
\item \href{#method-allByParticipantWithHttpInfo}{\code{ActivityEventApi$allByParticipantWithHttpInfo()}}
\item \href{#method-allByResearcher}{\code{ActivityEventApi$allByResearcher()}}
\item \href{#method-allByResearcherWithHttpInfo}{\code{ActivityEventApi$allByResearcherWithHttpInfo()}}
\item \href{#method-allByStudy}{\code{ActivityEventApi$allByStudy()}}
\item \href{#method-allByStudyWithHttpInfo}{\code{ActivityEventApi$allByStudyWithHttpInfo()}}
\item \href{#method-create}{\code{ActivityEventApi$create()}}
\item \href{#method-createWithHttpInfo}{\code{ActivityEventApi$createWithHttpInfo()}}
\item \href{#method-delete}{\code{ActivityEventApi$delete()}}
\item \href{#method-deleteWithHttpInfo}{\code{ActivityEventApi$deleteWithHttpInfo()}}
\item \href{#method-clone}{\code{ActivityEventApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByParticipant"></a>}}
\if{latex}{\out{\hypertarget{method-allByParticipant}{}}}
\subsection{Method \code{allByParticipant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventApi$allByParticipant(
  participant.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByParticipantWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allByParticipantWithHttpInfo}{}}}
\subsection{Method \code{allByParticipantWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventApi$allByParticipantWithHttpInfo(
  participant.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByResearcher"></a>}}
\if{latex}{\out{\hypertarget{method-allByResearcher}{}}}
\subsection{Method \code{allByResearcher()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventApi$allByResearcher(
  researcher.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByResearcherWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allByResearcherWithHttpInfo}{}}}
\subsection{Method \code{allByResearcherWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventApi$allByResearcherWithHttpInfo(
  researcher.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByStudy"></a>}}
\if{latex}{\out{\hypertarget{method-allByStudy}{}}}
\subsection{Method \code{allByStudy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventApi$allByStudy(
  study.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByStudyWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allByStudyWithHttpInfo}{}}}
\subsection{Method \code{allByStudyWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventApi$allByStudyWithHttpInfo(
  study.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create"></a>}}
\if{latex}{\out{\hypertarget{method-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventApi$create(participant.id, activity.event, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-createWithHttpInfo}{}}}
\subsection{Method \code{createWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventApi$createWithHttpInfo(participant.id, activity.event, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventApi$delete(
  participant.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deleteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-deleteWithHttpInfo}{}}}
\subsection{Method \code{deleteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventApi$deleteWithHttpInfo(
  participant.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
