% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_interval_legacy.R
\docType{class}
\name{DurationIntervalLegacy}
\alias{DurationIntervalLegacy}
\title{DurationIntervalLegacy}
\format{
An \code{R6Class} generator object
}
\description{
DurationIntervalLegacy Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{repeat_type}}{character [optional]}

\item{\code{date}}{numeric [optional]}

\item{\code{custom_times}}{list( object ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DurationIntervalLegacy$new()}}
\item \href{#method-toJSON}{\code{DurationIntervalLegacy$toJSON()}}
\item \href{#method-fromJSON}{\code{DurationIntervalLegacy$fromJSON()}}
\item \href{#method-toJSONString}{\code{DurationIntervalLegacy$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DurationIntervalLegacy$fromJSONString()}}
\item \href{#method-clone}{\code{DurationIntervalLegacy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationIntervalLegacy$new(
  repeat_type = NULL,
  date = NULL,
  custom_times = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationIntervalLegacy$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationIntervalLegacy$fromJSON(DurationIntervalLegacyJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationIntervalLegacy$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationIntervalLegacy$fromJSONString(DurationIntervalLegacyJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationIntervalLegacy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
