% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_attachment.R
\docType{class}
\name{DynamicAttachment}
\alias{DynamicAttachment}
\title{DynamicAttachment}
\format{
An \code{R6Class} generator object
}
\description{
DynamicAttachment Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{key}}{character [optional]}

\item{\code{from}}{character [optional]}

\item{\code{to}}{character [optional]}

\item{\code{triggers}}{list( object ) [optional]}

\item{\code{language}}{character [optional]}

\item{\code{contents}}{character [optional]}

\item{\code{requirements}}{list( object ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DynamicAttachment$new()}}
\item \href{#method-toJSON}{\code{DynamicAttachment$toJSON()}}
\item \href{#method-fromJSON}{\code{DynamicAttachment$fromJSON()}}
\item \href{#method-toJSONString}{\code{DynamicAttachment$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DynamicAttachment$fromJSONString()}}
\item \href{#method-clone}{\code{DynamicAttachment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicAttachment$new(
  key = NULL,
  from = NULL,
  to = NULL,
  triggers = NULL,
  language = NULL,
  contents = NULL,
  requirements = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicAttachment$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicAttachment$fromJSON(DynamicAttachmentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicAttachment$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicAttachment$fromJSONString(DynamicAttachmentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicAttachment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
