% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_api.R
\docType{class}
\name{ParticipantApi}
\alias{ParticipantApi}
\title{Participant operations}
\format{
An \code{R6Class} generator object
}
\description{
LAMP.Participant
}
\section{Methods}{

\describe{
\strong{ all } \emph{ Get the set of all participants. }
Get the set of all participants.

\itemize{
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ allByResearcher } \emph{ Get the set of all participants under a single researcher. }
Get the set of all participants under a single researcher.

\itemize{
\item \emph{ @param } researcher.id character
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ allByStudy } \emph{ Get the set of all participants in a single study. }
Get the set of all participants in a single study.

\itemize{
\item \emph{ @param } study.id character
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ create } \emph{ Create a new Participant for the given Study. }
Create a new Participant for the given Study.

\itemize{
\item \emph{ @param } study.id character
\item \emph{ @param } participant \link{Participant}


\item status code : 200 | 200 Success

\item return type : character 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ delete } \emph{ Delete a participant AND all owned data or event streams. }
Delete a participant AND all owned data or event streams.

\itemize{
\item \emph{ @param } participant.id character


\item status code : 200 | 200 Success

\item return type : character 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ update } \emph{ Update a Participant&#39;s settings. }
Update a Participant&#39;s settings.

\itemize{
\item \emph{ @param } participant.id character
\item \emph{ @param } participant \link{Participant}


\item status code : 200 | 200 Success

\item return type : character 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ view } \emph{ Get a single participant, by identifier. }
Get a single participant, by identifier.

\itemize{
\item \emph{ @param } participant.id character
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  all  ####################

library(LAMP)

#Get the set of all participants.
result <- LAMP.Participant$all()


####################  allByResearcher  ####################

library(LAMP)
var.researcher.id <- 'researcher.id_example' # character | 

#Get the set of all participants under a single researcher.
result <- LAMP.Participant$allByResearcher(var.researcher.id)


####################  allByStudy  ####################

library(LAMP)
var.study.id <- 'study.id_example' # character | 

#Get the set of all participants in a single study.
result <- LAMP.Participant$allByStudy(var.study.id)


####################  create  ####################

library(LAMP)
var.study.id <- 'study.id_example' # character | 
var.participant <- Participant$new() # Participant | 

#Create a new Participant for the given Study.
result <- LAMP.Participant$create(var.study.id, var.participant)


####################  delete  ####################

library(LAMP)
var.participant.id <- 'participant.id_example' # character | 

#Delete a participant AND all owned data or event streams.
result <- LAMP.Participant$delete(var.participant.id)


####################  update  ####################

library(LAMP)
var.participant.id <- 'participant.id_example' # character | 
var.participant <- Participant$new() # Participant | 

#Update a Participant's settings.
result <- LAMP.Participant$update(var.participant.id, var.participant)


####################  view  ####################

library(LAMP)
var.participant.id <- 'participant.id_example' # character | 

#Get a single participant, by identifier.
result <- LAMP.Participant$view(var.participant.id)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ParticipantApi$new()}}
\item \href{#method-all}{\code{ParticipantApi$all()}}
\item \href{#method-allWithHttpInfo}{\code{ParticipantApi$allWithHttpInfo()}}
\item \href{#method-allByResearcher}{\code{ParticipantApi$allByResearcher()}}
\item \href{#method-allByResearcherWithHttpInfo}{\code{ParticipantApi$allByResearcherWithHttpInfo()}}
\item \href{#method-allByStudy}{\code{ParticipantApi$allByStudy()}}
\item \href{#method-allByStudyWithHttpInfo}{\code{ParticipantApi$allByStudyWithHttpInfo()}}
\item \href{#method-create}{\code{ParticipantApi$create()}}
\item \href{#method-createWithHttpInfo}{\code{ParticipantApi$createWithHttpInfo()}}
\item \href{#method-delete}{\code{ParticipantApi$delete()}}
\item \href{#method-deleteWithHttpInfo}{\code{ParticipantApi$deleteWithHttpInfo()}}
\item \href{#method-update}{\code{ParticipantApi$update()}}
\item \href{#method-updateWithHttpInfo}{\code{ParticipantApi$updateWithHttpInfo()}}
\item \href{#method-view}{\code{ParticipantApi$view()}}
\item \href{#method-viewWithHttpInfo}{\code{ParticipantApi$viewWithHttpInfo()}}
\item \href{#method-clone}{\code{ParticipantApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all"></a>}}
\if{latex}{\out{\hypertarget{method-all}{}}}
\subsection{Method \code{all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$all(transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allWithHttpInfo}{}}}
\subsection{Method \code{allWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$allWithHttpInfo(transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByResearcher"></a>}}
\if{latex}{\out{\hypertarget{method-allByResearcher}{}}}
\subsection{Method \code{allByResearcher()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$allByResearcher(researcher.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByResearcherWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allByResearcherWithHttpInfo}{}}}
\subsection{Method \code{allByResearcherWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$allByResearcherWithHttpInfo(
  researcher.id,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByStudy"></a>}}
\if{latex}{\out{\hypertarget{method-allByStudy}{}}}
\subsection{Method \code{allByStudy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$allByStudy(study.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByStudyWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allByStudyWithHttpInfo}{}}}
\subsection{Method \code{allByStudyWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$allByStudyWithHttpInfo(study.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create"></a>}}
\if{latex}{\out{\hypertarget{method-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$create(study.id, participant, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-createWithHttpInfo}{}}}
\subsection{Method \code{createWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$createWithHttpInfo(study.id, participant, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$delete(participant.id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deleteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-deleteWithHttpInfo}{}}}
\subsection{Method \code{deleteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$deleteWithHttpInfo(participant.id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$update(participant.id, participant, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-updateWithHttpInfo}{}}}
\subsection{Method \code{updateWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$updateWithHttpInfo(participant.id, participant, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-view"></a>}}
\if{latex}{\out{\hypertarget{method-view}{}}}
\subsection{Method \code{view()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$view(participant.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-viewWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-viewWithHttpInfo}{}}}
\subsection{Method \code{viewWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$viewWithHttpInfo(participant.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParticipantApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
