% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_api.R
\docType{class}
\name{SensorApi}
\alias{SensorApi}
\title{Sensor operations}
\format{
An \code{R6Class} generator object
}
\description{
LAMP.Sensor
}
\section{Methods}{

\describe{
\strong{ all } \emph{ Get the set of all sensors. }
Get the set of all sensors.

\itemize{
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ allByParticipant } \emph{ Get all sensors for a participant. }
Get the set of all sensors available to a participant, by participant  identifier.

\itemize{
\item \emph{ @param } participant.id character
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ allByResearcher } \emph{ Get all sensors for a researcher. }
Get the set of all sensors available to participants of any study conducted  by a researcher, by researcher identifier.

\itemize{
\item \emph{ @param } researcher.id character
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ allByStudy } \emph{ View all sensors in a study. }
Get the set of all sensors available to participants of a single  study, by study identifier.

\itemize{
\item \emph{ @param } study.id character
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ create } \emph{ Create a new Sensor under the given Study. }
Create a new Sensor under the given Study.

\itemize{
\item \emph{ @param } study.id character
\item \emph{ @param } sensor \link{Sensor}


\item status code : 200 | 200 Success

\item return type : character 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ delete } \emph{ Delete a Sensor. }
Delete a Sensor.

\itemize{
\item \emph{ @param } sensor.id character


\item status code : 200 | 200 Success

\item return type : character 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ update } \emph{ Update an Sensor&#39;s settings. }
Update an Sensor&#39;s settings.

\itemize{
\item \emph{ @param } sensor.id character
\item \emph{ @param } sensor \link{Sensor}


\item status code : 200 | 200 Success

\item return type : character 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ view } \emph{ Get a single sensor, by identifier. }
Get a single sensor, by identifier.

\itemize{
\item \emph{ @param } sensor.id character
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  all  ####################

library(LAMP)

#Get the set of all sensors.
result <- LAMP.Sensor$all()


####################  allByParticipant  ####################

library(LAMP)
var.participant.id <- 'participant.id_example' # character | 

#Get all sensors for a participant.
result <- LAMP.Sensor$allByParticipant(var.participant.id)


####################  allByResearcher  ####################

library(LAMP)
var.researcher.id <- 'researcher.id_example' # character | 

#Get all sensors for a researcher.
result <- LAMP.Sensor$allByResearcher(var.researcher.id)


####################  allByStudy  ####################

library(LAMP)
var.study.id <- 'study.id_example' # character | 

#View all sensors in a study.
result <- LAMP.Sensor$allByStudy(var.study.id)


####################  create  ####################

library(LAMP)
var.study.id <- 'study.id_example' # character | 
var.sensor <- Sensor$new() # Sensor | 

#Create a new Sensor under the given Study.
result <- LAMP.Sensor$create(var.study.id, var.sensor)


####################  delete  ####################

library(LAMP)
var.sensor.id <- 'sensor.id_example' # character | 

#Delete a Sensor.
result <- LAMP.Sensor$delete(var.sensor.id)


####################  update  ####################

library(LAMP)
var.sensor.id <- 'sensor.id_example' # character | 
var.sensor <- Sensor$new() # Sensor | 

#Update an Sensor's settings.
result <- LAMP.Sensor$update(var.sensor.id, var.sensor)


####################  view  ####################

library(LAMP)
var.sensor.id <- 'sensor.id_example' # character | 

#Get a single sensor, by identifier.
result <- LAMP.Sensor$view(var.sensor.id)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SensorApi$new()}}
\item \href{#method-all}{\code{SensorApi$all()}}
\item \href{#method-allWithHttpInfo}{\code{SensorApi$allWithHttpInfo()}}
\item \href{#method-allByParticipant}{\code{SensorApi$allByParticipant()}}
\item \href{#method-allByParticipantWithHttpInfo}{\code{SensorApi$allByParticipantWithHttpInfo()}}
\item \href{#method-allByResearcher}{\code{SensorApi$allByResearcher()}}
\item \href{#method-allByResearcherWithHttpInfo}{\code{SensorApi$allByResearcherWithHttpInfo()}}
\item \href{#method-allByStudy}{\code{SensorApi$allByStudy()}}
\item \href{#method-allByStudyWithHttpInfo}{\code{SensorApi$allByStudyWithHttpInfo()}}
\item \href{#method-create}{\code{SensorApi$create()}}
\item \href{#method-createWithHttpInfo}{\code{SensorApi$createWithHttpInfo()}}
\item \href{#method-delete}{\code{SensorApi$delete()}}
\item \href{#method-deleteWithHttpInfo}{\code{SensorApi$deleteWithHttpInfo()}}
\item \href{#method-update}{\code{SensorApi$update()}}
\item \href{#method-updateWithHttpInfo}{\code{SensorApi$updateWithHttpInfo()}}
\item \href{#method-view}{\code{SensorApi$view()}}
\item \href{#method-viewWithHttpInfo}{\code{SensorApi$viewWithHttpInfo()}}
\item \href{#method-clone}{\code{SensorApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all"></a>}}
\if{latex}{\out{\hypertarget{method-all}{}}}
\subsection{Method \code{all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$all(transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allWithHttpInfo}{}}}
\subsection{Method \code{allWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$allWithHttpInfo(transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByParticipant"></a>}}
\if{latex}{\out{\hypertarget{method-allByParticipant}{}}}
\subsection{Method \code{allByParticipant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$allByParticipant(participant.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByParticipantWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allByParticipantWithHttpInfo}{}}}
\subsection{Method \code{allByParticipantWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$allByParticipantWithHttpInfo(participant.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByResearcher"></a>}}
\if{latex}{\out{\hypertarget{method-allByResearcher}{}}}
\subsection{Method \code{allByResearcher()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$allByResearcher(researcher.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByResearcherWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allByResearcherWithHttpInfo}{}}}
\subsection{Method \code{allByResearcherWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$allByResearcherWithHttpInfo(researcher.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByStudy"></a>}}
\if{latex}{\out{\hypertarget{method-allByStudy}{}}}
\subsection{Method \code{allByStudy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$allByStudy(study.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByStudyWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allByStudyWithHttpInfo}{}}}
\subsection{Method \code{allByStudyWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$allByStudyWithHttpInfo(study.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create"></a>}}
\if{latex}{\out{\hypertarget{method-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$create(study.id, sensor, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-createWithHttpInfo}{}}}
\subsection{Method \code{createWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$createWithHttpInfo(study.id, sensor, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$delete(sensor.id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deleteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-deleteWithHttpInfo}{}}}
\subsection{Method \code{deleteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$deleteWithHttpInfo(sensor.id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$update(sensor.id, sensor, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-updateWithHttpInfo}{}}}
\subsection{Method \code{updateWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$updateWithHttpInfo(sensor.id, sensor, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-view"></a>}}
\if{latex}{\out{\hypertarget{method-view}{}}}
\subsection{Method \code{view()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$view(sensor.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-viewWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-viewWithHttpInfo}{}}}
\subsection{Method \code{viewWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$viewWithHttpInfo(sensor.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
