% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_event_api.R
\docType{class}
\name{SensorEventApi}
\alias{SensorEventApi}
\title{SensorEvent operations}
\format{
An \code{R6Class} generator object
}
\description{
LAMP.SensorEvent
}
\section{Methods}{

\describe{
\strong{ allByParticipant } \emph{ Get all sensor events for a participant. }
Get the set of all sensor events produced by the given participant.

\itemize{
\item \emph{ @param } participant.id character
\item \emph{ @param } origin character
\item \emph{ @param } from numeric
\item \emph{ @param } to numeric
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ allByResearcher } \emph{ Get all sensor events for a researcher by participant. }
Get the set of all sensor events produced by participants of any  study conducted by a researcher, by researcher identifier.

\itemize{
\item \emph{ @param } researcher.id character
\item \emph{ @param } origin character
\item \emph{ @param } from numeric
\item \emph{ @param } to numeric
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ allByStudy } \emph{ Get all sensor events for a study by participant. }
Get the set of all sensor events produced by participants of a  single study, by study identifier.

\itemize{
\item \emph{ @param } study.id character
\item \emph{ @param } origin character
\item \emph{ @param } from numeric
\item \emph{ @param } to numeric
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ create } \emph{ Create a new SensorEvent for the given Participant. }
Create a new SensorEvent for the given Participant.

\itemize{
\item \emph{ @param } participant.id character
\item \emph{ @param } sensor.event \link{SensorEvent}


\item status code : 200 | 200 Success

\item return type : character 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ delete } \emph{ Delete a sensor event. }
Delete a sensor event.

\itemize{
\item \emph{ @param } participant.id character
\item \emph{ @param } origin character
\item \emph{ @param } from numeric
\item \emph{ @param } to numeric


\item status code : 200 | 200 Success

\item return type : character 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  allByParticipant  ####################

library(LAMP)
var.participant.id <- 'participant.id_example' # character | 
var.origin <- 'origin_example' # character | 
var.from <- 3.4 # numeric | 
var.to <- 3.4 # numeric | 

#Get all sensor events for a participant.
result <- LAMP.SensorEvent$allByParticipant(var.participant.id, 
                                           origin=var.origin, 
                                           from=var.from, 
                                           to=var.to)


####################  allByResearcher  ####################

library(LAMP)
var.researcher.id <- 'researcher.id_example' # character | 
var.origin <- 'origin_example' # character | 
var.from <- 3.4 # numeric | 
var.to <- 3.4 # numeric | 

#Get all sensor events for a researcher by participant.
result <- LAMP.SensorEvent$allByResearcher(var.researcher.id, 
                                          origin=var.origin, 
                                          from=var.from, 
                                          to=var.to)


####################  allByStudy  ####################

library(LAMP)
var.study.id <- 'study.id_example' # character | 
var.origin <- 'origin_example' # character | 
var.from <- 3.4 # numeric | 
var.to <- 3.4 # numeric | 

#Get all sensor events for a study by participant.
result <- LAMP.SensorEvent$allByStudy(var.study.id, origin=var.origin, from=var.from, to=var.to)


####################  create  ####################

library(LAMP)
var.participant.id <- 'participant.id_example' # character | 
var.sensor.event <- SensorEvent$new() # SensorEvent | 

#Create a new SensorEvent for the given Participant.
result <- LAMP.SensorEvent$create(var.participant.id, var.sensor.event)


####################  delete  ####################

library(LAMP)
var.participant.id <- 'participant.id_example' # character | 
var.origin <- 'origin_example' # character | 
var.from <- 3.4 # numeric | 
var.to <- 3.4 # numeric | 

#Delete a sensor event.
result <- LAMP.SensorEvent$delete(var.participant.id, origin=var.origin, from=var.from, to=var.to)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SensorEventApi$new()}}
\item \href{#method-allByParticipant}{\code{SensorEventApi$allByParticipant()}}
\item \href{#method-allByParticipantWithHttpInfo}{\code{SensorEventApi$allByParticipantWithHttpInfo()}}
\item \href{#method-allByResearcher}{\code{SensorEventApi$allByResearcher()}}
\item \href{#method-allByResearcherWithHttpInfo}{\code{SensorEventApi$allByResearcherWithHttpInfo()}}
\item \href{#method-allByStudy}{\code{SensorEventApi$allByStudy()}}
\item \href{#method-allByStudyWithHttpInfo}{\code{SensorEventApi$allByStudyWithHttpInfo()}}
\item \href{#method-create}{\code{SensorEventApi$create()}}
\item \href{#method-createWithHttpInfo}{\code{SensorEventApi$createWithHttpInfo()}}
\item \href{#method-delete}{\code{SensorEventApi$delete()}}
\item \href{#method-deleteWithHttpInfo}{\code{SensorEventApi$deleteWithHttpInfo()}}
\item \href{#method-clone}{\code{SensorEventApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorEventApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByParticipant"></a>}}
\if{latex}{\out{\hypertarget{method-allByParticipant}{}}}
\subsection{Method \code{allByParticipant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorEventApi$allByParticipant(
  participant.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByParticipantWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allByParticipantWithHttpInfo}{}}}
\subsection{Method \code{allByParticipantWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorEventApi$allByParticipantWithHttpInfo(
  participant.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByResearcher"></a>}}
\if{latex}{\out{\hypertarget{method-allByResearcher}{}}}
\subsection{Method \code{allByResearcher()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorEventApi$allByResearcher(
  researcher.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByResearcherWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allByResearcherWithHttpInfo}{}}}
\subsection{Method \code{allByResearcherWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorEventApi$allByResearcherWithHttpInfo(
  researcher.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByStudy"></a>}}
\if{latex}{\out{\hypertarget{method-allByStudy}{}}}
\subsection{Method \code{allByStudy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorEventApi$allByStudy(
  study.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByStudyWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allByStudyWithHttpInfo}{}}}
\subsection{Method \code{allByStudyWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorEventApi$allByStudyWithHttpInfo(
  study.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  transform = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create"></a>}}
\if{latex}{\out{\hypertarget{method-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorEventApi$create(participant.id, sensor.event, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-createWithHttpInfo}{}}}
\subsection{Method \code{createWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorEventApi$createWithHttpInfo(participant.id, sensor.event, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorEventApi$delete(
  participant.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deleteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-deleteWithHttpInfo}{}}}
\subsection{Method \code{deleteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorEventApi$deleteWithHttpInfo(
  participant.id,
  origin = NULL,
  from = NULL,
  to = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorEventApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
