% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_api.R
\docType{class}
\name{StudyApi}
\alias{StudyApi}
\title{Study operations}
\format{
An \code{R6Class} generator object
}
\description{
LAMP.Study
}
\section{Methods}{

\describe{
\strong{ all } \emph{ Get the set of all studies. }
Get the set of all studies.

\itemize{
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ allByResearcher } \emph{ Get the set of studies for a single researcher. }
Get the set of studies for a single researcher.

\itemize{
\item \emph{ @param } researcher.id character
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ create } \emph{ Create a new Study for the given Researcher. }
Create a new Study for the given Researcher.

\itemize{
\item \emph{ @param } researcher.id character
\item \emph{ @param } study \link{Study}


\item status code : 200 | 200 Success

\item return type : character 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ delete } \emph{ Delete a study. }
Delete a study.

\itemize{
\item \emph{ @param } study.id character


\item status code : 200 | 200 Success

\item return type : character 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ update } \emph{ Update the study. }
Update the study.

\itemize{
\item \emph{ @param } study.id character
\item \emph{ @param } study \link{Study}


\item status code : 200 | 200 Success

\item return type : character 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ view } \emph{ Get a single study, by identifier. }
Get a single study, by identifier.

\itemize{
\item \emph{ @param } study.id character
\item \emph{ @param } transform character


\item status code : 200 | 200 Success

\item return type : array[object] 
\item response headers :

\tabular{ll}{
}
\item status code : 400 | 400 Bad Request

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 403 | 403 Authorization Failed

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | 404 Not Found

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | 500 Internal Error

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  all  ####################

library(LAMP)

#Get the set of all studies.
result <- LAMP.Study$all()


####################  allByResearcher  ####################

library(LAMP)
var.researcher.id <- 'researcher.id_example' # character | 

#Get the set of studies for a single researcher.
result <- LAMP.Study$allByResearcher(var.researcher.id)


####################  create  ####################

library(LAMP)
var.researcher.id <- 'researcher.id_example' # character | 
var.study <- Study$new() # Study | 

#Create a new Study for the given Researcher.
result <- LAMP.Study$create(var.researcher.id, var.study)


####################  delete  ####################

library(LAMP)
var.study.id <- 'study.id_example' # character | 

#Delete a study.
result <- LAMP.Study$delete(var.study.id)


####################  update  ####################

library(LAMP)
var.study.id <- 'study.id_example' # character | 
var.study <- Study$new() # Study | 

#Update the study.
result <- LAMP.Study$update(var.study.id, var.study)


####################  view  ####################

library(LAMP)
var.study.id <- 'study.id_example' # character | 

#Get a single study, by identifier.
result <- LAMP.Study$view(var.study.id)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StudyApi$new()}}
\item \href{#method-all}{\code{StudyApi$all()}}
\item \href{#method-allWithHttpInfo}{\code{StudyApi$allWithHttpInfo()}}
\item \href{#method-allByResearcher}{\code{StudyApi$allByResearcher()}}
\item \href{#method-allByResearcherWithHttpInfo}{\code{StudyApi$allByResearcherWithHttpInfo()}}
\item \href{#method-create}{\code{StudyApi$create()}}
\item \href{#method-createWithHttpInfo}{\code{StudyApi$createWithHttpInfo()}}
\item \href{#method-delete}{\code{StudyApi$delete()}}
\item \href{#method-deleteWithHttpInfo}{\code{StudyApi$deleteWithHttpInfo()}}
\item \href{#method-update}{\code{StudyApi$update()}}
\item \href{#method-updateWithHttpInfo}{\code{StudyApi$updateWithHttpInfo()}}
\item \href{#method-view}{\code{StudyApi$view()}}
\item \href{#method-viewWithHttpInfo}{\code{StudyApi$viewWithHttpInfo()}}
\item \href{#method-clone}{\code{StudyApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all"></a>}}
\if{latex}{\out{\hypertarget{method-all}{}}}
\subsection{Method \code{all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$all(transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allWithHttpInfo}{}}}
\subsection{Method \code{allWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$allWithHttpInfo(transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByResearcher"></a>}}
\if{latex}{\out{\hypertarget{method-allByResearcher}{}}}
\subsection{Method \code{allByResearcher()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$allByResearcher(researcher.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allByResearcherWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-allByResearcherWithHttpInfo}{}}}
\subsection{Method \code{allByResearcherWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$allByResearcherWithHttpInfo(researcher.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create"></a>}}
\if{latex}{\out{\hypertarget{method-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$create(researcher.id, study, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-createWithHttpInfo}{}}}
\subsection{Method \code{createWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$createWithHttpInfo(researcher.id, study, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$delete(study.id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deleteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-deleteWithHttpInfo}{}}}
\subsection{Method \code{deleteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$deleteWithHttpInfo(study.id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$update(study.id, study, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-updateWithHttpInfo}{}}}
\subsection{Method \code{updateWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$updateWithHttpInfo(study.id, study, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-view"></a>}}
\if{latex}{\out{\hypertarget{method-view}{}}}
\subsection{Method \code{view()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$view(study.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-viewWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-viewWithHttpInfo}{}}}
\subsection{Method \code{viewWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$viewWithHttpInfo(study.id, transform = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
